refactor(view): lift the queue float into a module of its own
`view.lua` is the most-read and most-changed file here, and the float over the
queue is the second of the two parts of it that own none of the view's mutable
state. The queue is a module-level singleton elsewhere; a float over it needs a
buffer, a window and keys, and nothing else.

The float's window handle stays on the view, behind `hold_queue_float` and
`release_queue_float`. Closing a float on submit is a rule about the view's
windows -- submit runs from either pane as readily as from the float -- and
moving the handle out would split one rule across two modules to save two
accessors.

The jump from a queued entry into the diff stays too, despite reading as part of
the float: it reaches the file list, both layouts' renders, expansion state and
both panes, and leaving it behind is precisely why the moved file is small.

The view is handed in rather than required, as `keymaps.lua` is handed it, which
keeps the new module out of the cycle `view` and `annotate` sit in.

Closes #81
